requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
mysqlclient
python-dotenv